No longer being wishy-washy about what things convert into. Cell arrays -> tuples Scalars -> scalars Nonscalars -> lists Numeric types will probably convert to numpy arrays in the future, once I'm done reading the numpy docs.
The colon operator operates somewhat correctly, as does the end operator.
If you index with just a colon, you get the uberslice. The end operator produces
len(obj)-1, which will be a python integer. If you use the colon operator and at
least one of the arguments is a python object, the result will be a slice.
numpy arrays have differing behavior depending on whether they're indexed with
a slice or a list/array, so it might be surprising that x{0:9} is different than
x{0:end}. We may want to think about how to resolve this.

Adding the 'types' package. Different python types may benefit from having
different behavior in MATLAB. To support this without putting special cases
all over the place, just create a class in py.types that extends py.types.object.
Make sure the class has the same name as the python type to be specialized.
The most specific class found in the python type's MRO list will be used.

If you do not assign the result of py.import to something, it will automatically
assign it for you. So you can do:
py.import('numpy');
x = numpy.arange(10);
etc. There seems to be an unresolved bug somewhere. Importing numpy will sometimes
crash MATLAB. Jealousy, I think. In any case, py.import will check the modules dict
before doing the import, so if it works once it should continue to work until
python is cleared. Maybe.

Normalizing my behavior of builtin shortcuts. py.tuple, list, and dict now
work /exactly/ like the python builtins, because that's what they are. Particularly
in the case of dict, this isn't very friendly. So I have also included the
shortcuts T, L, and D which operate on varargs.
py.T(1,2,3) -> (1.0, 2.0, 3.0)
py.L(1,2,3) -> [1.0, 2.0, 3.0]
py.D('a', 1, 'b', 2) -> {'a':1.0, 'b':2.0}