Alright, what the hell have I done tonight? I'm tired, so we'll see what I can recall.
There is now a BasePyObject type above object. Currently the only other
direct subclass is 'null', for null pointers. Any special handling of those
can go there without having to worry about dropping into one of object's
methods.

I've changed colon. It first generates a standard range. However,
MATLAB will only do "for i=..." iteration over MATLAB matrices, so
the range will attempt to convert to int. Or maybe double.

py.builtins will do a few cute things depending on number of outputs.
If no inputs and one output, just produces th builtins dict.
If multiple inputs and one output, they'll be returned as a tuple.
If multiple inputs and multiple outputs, the outputs will be paired
to the inputs.
If any inputs and no outputs, the objects will be assigned into the caller's
workspace with appropriate names. Items with underscore names will not
be assigned in.
If no inputs and no outputs, ALL builtins will be assigned in to the caller's
workspace (except the __ names, of course).

Concatenating generic objects (along any dimension) will generate
a list. The list will be extended with iterables or appended with
non-iterables. This might not be the most useful syntax in the world,
but it keeps the MATLAB handles scalar while doing something meaningful.
Concatenation of numpy arrays will probably do something more useful
in the future.