I figured out how to properly override/supplement mex's compiler flags, so I've built this into the Makefile and removed the custom mexopts.sh. Now mex will use your default mexopts, or you can provide one by saying
make MEXFLAGS="-f /path/to/myspecialopts.sh"

Since we're not trying to use a specific options file, It's entirely
possible that this will now work in cygwin, so I'll have
to boot Windows one of these days and give it a whirl.
MATLAB for Windows uses a 'mexopts.bat' instead of a bash script,
so it probably needs some changes (use %s instead of $s for shell vars).